fixed types, added pool to stack

+++ execute.h
@@ -1,3 +1,3 @@
 #include "types.h"
 
-int executeWord(int index, ForthToken* tokens, int tokenLength);
+int executeWord(int index, ForthToken* tokens, size_t tokenLength);
